test: keep ACP manager helpers within task boundaries
